refactoring

+++ backend/de.metas.banking/de.metas.banking.base/src/main/java/de/metas/banking/payment/impl/PaymentRequestBL.java
@@ -109,7 +109,7 @@ public class PaymentRequestBL implements IPaymentRequestBL
 		{
 			// task 09698: don't apply more than the amount which is actually still open, even if the paymentRequest's amount is bigger.
 			final boolean creditMemoAdjusted = true;
-			final BigDecimal openAmt = allocationDAO.retrieveOpenAmtInInvoiceCurrency(invoice, creditMemoAdjusted);
+			final BigDecimal openAmt = allocationDAO.retrieveOpenAmtInInvoiceCurrency(invoice, creditMemoAdjusted).toBigDecimal();
 			final BigDecimal payAmt = requestAmount.min(openAmt);
 			paySelectionLine.setPayAmt(payAmt);
